% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{beta_addin}
\alias{beta_addin}
\title{Beta distribution gadget}
\usage{
beta_addin()
}
\value{
A numeric vector containing the chosen shape parameter values.
}
\description{
This gadget lets you choose the parameters of a Beta
  distribution easily. It supports two parameterization options: the classic
  one with shape parameters and a more intuitive one using the expectation
  and precision as proposed by Ferrari and Cribari-Neto (2004).
}
\references{
Ferrari, S., & Cribari-Neto, F. (2004). Beta regression for
  modelling rates and proportions. Journal of Applied Statistics.
}

