% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinydensR.R
\docType{package}
\name{tinydensR-package}
\alias{tinydensR-package}
\alias{tinydensR}
\title{tinydensR: RStudio add-ins for interactively visualizing probability distributions}
\description{
A set of RStudio add-ins for playing with distribution parameters
  and visualizing the resulting probability density and mass functions.
}
