% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-discrete.R
\name{univariate_discrete_addin}
\alias{univariate_discrete_addin}
\title{Univariate discrete distribution gadget}
\usage{
univariate_discrete_addin()
}
\value{
A named vector containing the chosen parameter value(s).
}
\description{
This gadget lets you choose the parameters of a univariate,
discrete distribution easily.
}
